REFACTOR: Rename utility to units
